Move PlayerJoinListener and PlayerQuitListener to listeners package.
Extract inner classes from Pathinator.java into dedicated files alongside PlayerMoveListener. No behavioral change.
